refactor: split security monolith into modules, fix auth bugs
Split tet/security/authentication.py (1457 lines) into focused modules:
- config.py: dataclasses, constants, protocols, enums
- models.py: SQLAlchemy model mixins (TokenMixin, MFA mixin)
- policy.py: TokenAuthenticationPolicy (removed JWTCookieAuthenticationPolicy)
- tokens.py: TetTokenService (JWT + long-term token management)
- auth.py: TetAuthService (cookies, password, breach check)
- mfa.py: TetMultiFactorAuthenticationService (TOTP setup/verify/challenge)
- views.py: AuthViews (all HTTP endpoints)
- authentication.py: slim directive + includeme + backward-compat re-exports

Security fixes:
- Cookie defaults: secure=True, httponly=True (were both False)
- Fix shared mutable JWT claims via dataclasses.replace()
- Catch jwt.InvalidTokenError instead of just ExpiredSignatureError
- Add timeout=5 to breach API call with graceful degradation
- Return refresh_token in login response body (not just cookie)
- Accept refresh_token from request body in refresh endpoint
- Store route_prefix on registry at config time (not fragile runtime lookup)

Other improvements:
- Remove JWTCookieAuthenticationPolicy entirely
- Use keyword-only arguments on service methods
- Move auth deps to optional extras group in setup.py
- Add pytest-cov, update CI with coverage summary
- Add tests for JWT edge cases, refresh token, breach API timeout
